with todos completed

+++ cloud-project/src/main/java/it/unipi/hadoop/LetterFrequencyCombiner.java
@@ -1,18 +1,18 @@
 package it.unipi.hadoop;
 
 import java.io.IOException;
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Reducer;
 
-public class LetterFrequencyCombiner extends Reducer<Text, IntWritable, Text, IntWritable> {
+public class LetterFrequencyCombiner extends Reducer<Text, LongWritable, Text, LongWritable> {
 
     @Override
-    public void reduce(Text key, Iterable<IntWritable> values, Context context) throws IOException, InterruptedException {
+    public void reduce(Text key, Iterable<LongWritable> values, Context context) throws IOException, InterruptedException {
         int sum = 0;
-        for (IntWritable val : values) {
+        for (LongWritable val : values) {
             sum += val.get();
         }
-        context.write(key, new IntWritable(sum));
+        context.write(key, new LongWritable(sum));
     }
 }
