+++ cloud-project/src/main/java/it/unipi/hadoop/LetterFrequencyReducer.java
@@ -2,21 +2,21 @@ package it.unipi.hadoop;
 
 import java.io.IOException;
 
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Reducer;
 
-public class LetterFrequencyReducer extends Reducer<Text, IntWritable, Text, LongWritable> {
+public class LetterFrequencyReducer extends Reducer<Text, LongWritable, Text, LongWritable> {
 
     public enum Counters {
         TOTAL_LETTERS
     }
 
     @Override
-    public void reduce(Text key, Iterable<IntWritable> values, Context context) throws IOException, InterruptedException {
+    public void reduce(Text key, Iterable<LongWritable> values, Context context) throws IOException, InterruptedException {
         long sum = 0;
-        for (IntWritable val : values) {
+        for (LongWritable val : values) {
             sum += val.get();
         }
         context.getCounter(Counters.TOTAL_LETTERS).increment(sum);
